refactor: Channel, ReadStatus, Message 도메인 서비스 계층 JPA 리팩토링

+++ 2-sprint-mission/src/main/java/com/sprint/mission/discodeit/service/basic/BasicChannelService.java
@@ -4,29 +4,35 @@ import com.sprint.mission.discodeit.dto.ChannelDto;
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.ChannelType;
 import com.sprint.mission.discodeit.entity.ReadStatus;
+import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.entity.base.BaseEntity;
 import com.sprint.mission.discodeit.exception.BusinessException;
 import com.sprint.mission.discodeit.exception.ErrorCode;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.MessageRepository;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
+import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class BasicChannelService implements ChannelService {
 
   private final ChannelRepository channelRepository;
   private final ReadStatusRepository readStatusRepository;
   private final MessageRepository messageRepository;
+  private final UserRepository userRepository;
 
   @Override
+  @Transactional
   public ChannelDto.Response createPublicChannel(ChannelDto.CreatePublicRequest request) {
     Channel channel = request.toEntity();
     channelRepository.save(channel);
@@ -35,39 +41,41 @@ public class BasicChannelService implements ChannelService {
   }
 
   @Override
+  @Transactional
   public ChannelDto.Response createPrivateChannel(ChannelDto.CreatePrivateRequest request) {
     Channel channel = request.toEntity();
     channelRepository.save(channel);
+    if (request.participantIds() != null && !request.participantIds().isEmpty()) {
+      List<User> participants = userRepository.findAllById(request.participantIds());
 
-    // 채널 참여자 ReadStatus 생성
-    if (channel.getMemberIds() != null) {
-      channel.getMemberIds().forEach(userId -> {
-        ReadStatus readStatus = ReadStatus.builder()
-            .userId(userId)
-            .channelId(channel.getId())
-            .build();
-        readStatusRepository.save(readStatus);
-      });
+      List<ReadStatus> readStatuses = participants.stream()
+          .map(user -> ReadStatus.builder()
+              .user(user)
+              .channel(channel)
+              .lastReadAt(channel.getCreatedAt())
+              .build())
+          .toList();
+
+      readStatusRepository.saveAll(readStatuses);
     }
-
     return toResponse(channel);
   }
 
   // 공통 로직
   private ChannelDto.Response toResponse(Channel channel) {
     // 가장 최신 메시지
-    Instant lastMessageAt = messageRepository.findAll().stream()
-        .filter(m -> m.getChannelId().equals(channel.getId()))
+    Instant lastMessageAt = messageRepository.findTopByChannelIdOrderByCreatedAtDesc(
+            channel.getId())
         .map(BaseEntity::getCreatedAt)
-        .max(Instant::compareTo)
         .orElse(null);
 
-    // 채널 참여 멤버 목록
-    List<UUID> userIds = (channel.getType() == ChannelType.PRIVATE)
-        ? channel.getMemberIds()
-        : null;
-
-    return ChannelDto.Response.of(channel, lastMessageAt, userIds);
+    List<UUID> participantIds = null;
+    if (channel.getType() == ChannelType.PRIVATE) {
+      participantIds = readStatusRepository.findAllByChannelId(channel.getId()).stream()
+          .map(readStatus -> readStatus.getChannel().getId())
+          .toList();
+    }
+    return ChannelDto.Response.of(channel, lastMessageAt, participantIds);
   }
 
   @Override
@@ -80,19 +88,18 @@ public class BasicChannelService implements ChannelService {
 
   @Override
   public List<ChannelDto.Response> findAllByUserId(UUID userId) {
-    List<UUID> myChannelIds = readStatusRepository.findAll().stream()
-        .filter(rs -> rs.getUserId().equals(userId))
-        .map(ReadStatus::getChannelId)
+    List<UUID> myChannelIds = readStatusRepository.findAllByUserId(userId).stream()
+        .map(readStatus -> readStatus.getChannel().getId())
         .toList();
 
-    return channelRepository.findAll().stream()
-        .filter(channel -> channel.getType() == ChannelType.PUBLIC || myChannelIds.contains(
-            channel.getId()))
+    return channelRepository.findAllByTypeOrIdIn(ChannelType.PUBLIC, myChannelIds)
+        .stream()
         .map(this::toResponse)
         .toList();
   }
 
   @Override
+  @Transactional
   public ChannelDto.Response update(UUID id, ChannelDto.UpdateRequest request) {
     Channel channel = channelRepository.findById(id)
         .orElseThrow(() -> new BusinessException(ErrorCode.CHANNEL_NOT_FOUND));
@@ -106,21 +113,13 @@ public class BasicChannelService implements ChannelService {
   }
 
   @Override
+  @Transactional
   public void delete(UUID id) {
-    Channel channel = channelRepository.findById(id)
-        .orElseThrow(() -> new BusinessException(ErrorCode.CHANNEL_NOT_FOUND));
-
-    // 채널 내 메시지 삭제
-    messageRepository.findAll().stream()
-        .filter(m -> m.getChannelId().equals(id))
-        .toList()
-        .forEach(m -> messageRepository.deleteById(m.getId()));
-
-    // 채널 내 ReadStatus 삭제
-    readStatusRepository.findAll().stream()
-        .filter(rs -> rs.getChannelId().equals(id))
-        .toList()
-        .forEach(rs -> readStatusRepository.deleteById(rs.getId()));
+    if (!channelRepository.existsById(id)) {
+      throw new BusinessException(ErrorCode.CHANNEL_NOT_FOUND);
+    }
+    messageRepository.deleteByChannelId(id);
+    readStatusRepository.deleteByChannelId(id);
 
     channelRepository.deleteById(id);
   }
